fibonacci 3

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,19 +7,19 @@
  */
 int main(void)
 {
-	long long int i = 2, a = 1, b = 2;
-	long long int c = 0;
+	long int i = 2, a = 1, b = 2;
+	long int c = 0;
 
-	printf("%lld, %lld, ", a, b);
+	printf("%ld, %ld, ", a, b);
 	for (; i < 98; i++)
 	{
 		c = a + b;
 		a = b;
 		b = c;
 		if (i < 97)
-			printf("%lld, ", c);
+			printf("%ld, ", c);
 		if (i == 97)
-			printf("%lld\n", c);
+			printf("%ld\n", c);
 	}
 	return (0);
 }
